requirements.txt
Flask==2.3.2
tensorflow==2.18.0
pillow==10.3.0
numpy==1.23.5
pdf2image==1.16.3
opencv-python==4.6.0.66
